Update 2-what-is-github.md
I removed the subunit "The GitHub Flow," ironically for me it felt like it broke the flow of the unit, especially being at the very beginning. If it is in fact something someone must learn before they learn about the basic features of GitHub I feel like it should be a prerequisite to the course or be its own module. I also made some minor grammatical edits.