Update 2-manage-your-dependencies-github.md
When I reread this section I felt myself a bit confused by the unit name and why it didn't outwardly have something to do with Dependabot. So I retooled the dependabot section to come first that way we introduce how the dependency graph and the GitHub Advisory Database build into Dependabot.