Update 2-github-actions-automate-development-tasks.md
A couple of grammatical changes but I deleted the first two subunits. I felt like we did a good job priming the learners of what Actions could do so the first two units further went into that and it felt redundant and keeping people from learning about Actions which is what they came here for.