Update and rename 7-Projects-versus-Projects-Classic to 6-Projects-versus-Projects-Classic
Created a whole unit for the differences between Projects and Projects (Classic). I originally had this section in the 1st learning unit, but it felt out of place and cumbersome. After drawing some inspiration from Docs, I saw that they included their Projects (Classic) content at the end of their information and felt this was fitting.